+++ src/tables/mod.rs
@@ -15,6 +15,7 @@ pub mod cmap;
 pub mod colr;
 pub mod cpal;
 pub mod cvar;
+pub mod device;
 pub mod ebdt;
 pub mod ebsc;
 pub mod feature_variations;